requirements.txt
alembic==1.17.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
cachetools==6.2.1
celery==5.5.3
cryptography==46.0.3
distro==1.9.0
email-validator==2.3.0
fastapi==0.119.0
feedparser==6.0.12
filelock==3.20.0
gitdb==4.0.12
GitPython==3.1.45
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
kombu==5.5.4
lxml==6.0.2
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.13
narwhals==2.8.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
plotly==6.3.1
pluggy==1.6.0
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic-settings==2.11.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
smmap==5.0.2
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
tenacity==8.2.3
srsly==2.5.1
starlette==0.49.3
streamlit==1.50.0
tokenizers==0.22.1
toml==0.10.2
tqdm==4.67.1
typer==0.19.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.1
wcwidth==0.2.14
weasel==0.4.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
yarl==1.22.0
zxcvbn==4.5.0
